Hacks to get things building again and sync w/ latest work

+++ arch/arm64/include/asm/kvm_hyp.h
@@ -130,7 +130,7 @@ struct kvm_iommu_ops {
 	bool (*host_mmio_dabt_handler)(struct kvm_cpu_context *host_ctxt,
 				       phys_addr_t fault_pa, unsigned int len,
 				       bool is_write, int rd);
-	void (*host_stage2_set_owner)(phys_addr_t addr, size_t size, u8 owner_id);
+	void (*host_stage2_set_owner)(phys_addr_t addr, size_t size, u32 owner_id);
 	int (*host_stage2_adjust_mmio_range)(phys_addr_t addr, phys_addr_t *start,
 					     phys_addr_t *end);
 };
